requirements.txt
boto3
line-bot-sdk==1.12.1
openai